Satisfy source ownership guards
